requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
jeepney==0.9.0
pyee==13.0.0
python-ffmpeg==2.0.12
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
wakepy==0.10.2.post1
